% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAS.R
\name{cascheck}
\alias{cascheck}
\title{check if Cas number valid}
\usage{
cascheck(cas)
}
\arguments{
\item{cas}{string to check}
}
\value{
TRUE if string is a valid CAS, FALSE if not
}
\description{
This function uses the CAS structure see more info on https://www.cas.org/support/documentation/chemical-substances/faqs to check if a string is a valid CAS.
}
\keyword{cascheck}
