% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{checkSeqDb}
\alias{checkSeqDb}
\title{checkSeqDb}
\usage{
checkSeqDb(genomeVersion)
}
\arguments{
\item{genomeVersion}{String that denotes genome version. 
To unambigously select a BSgenome object, provide a string
that matches the end of the available genomes at: 
BSgenome::available.genomes().}
}
\value{
Returns a BSgenome object that uniquely matches the 
genomeVersion.
}
\description{
Given a string that denotes a genome version (e.g. hg19)
returns the BSgenome object matching the genome version 
that are available in BSgenome::available.genomes()
}
\examples{
checkSeqDb('hg19')
}
