% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_analysis.R
\name{createControlRegions}
\alias{createControlRegions}
\title{createControlRegions}
\usage{
createControlRegions(queryRegions)
}
\arguments{
\item{queryRegions}{GRanges object containing coordinates of input query
regions imported by the \code{\link{importBed}} function.}
}
\value{
GRanges object that contains the same number of regions as query regions
}
\description{
Given a GRanges object of query regions, create a background set of peaks that have
the same length distribution based on the flanking regions of the peaks.
}
\examples{
bed <- importBed('input.bed')
controlRegions <- createControlRegions(queryRegions=bed)

}

