% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{createOrthologousGeneSetList}
\alias{createOrthologousGeneSetList}
\title{createOrthologousMsigdbDataset
This function is deprecated. For functional enrichment 
analysis, use findEnrichedFunctions.}
\usage{
createOrthologousGeneSetList()
}
\description{
createOrthologousMsigdbDataset
This function is deprecated. For functional enrichment 
analysis, use findEnrichedFunctions.
}
