% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{createOrthologousMsigdbDataset}
\alias{createOrthologousMsigdbDataset}
\title{Create Orthologous Msigdb Dataset}
\usage{
createOrthologousMsigdbDataset(refMsigdbFilePath, refGenomeVersion = "hg19",
  targetGenomeVersion)
}
\arguments{
\item{refMsigdbFilePath}{Path to the file containing gene sets from MSIGDB. The gene ids must be
in Entrez format.}

\item{refGenomeVersion}{Genome version of the reference species. (default:hg19)}

\item{targetGenomeVersion}{Genome version of the target species. Available options are mm9, dm3, and ce6}
}
\value{
A list of vectors where each vector consists of a set of Entrez gene ids
}
\description{
MSIGDB curates gene sets for human but not for other species. This function is
used to create such gene sets for other species such as mouse, fly, and worm via
orthologous relationships to human genes.
}

