% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseSetUpFunctions.R
\name{deleteSampleDataFromDB}
\alias{deleteSampleDataFromDB}
\title{#' getContentSummaryFromDB
#' 
#' For a given database generated by \code{createDB} function, this function
#' reads the tables and their contents and provides a quick summary of what is
#' available in the database.
#' 
#' @param dbPath Path to an sqlite file generated by \code{createDB} function.
#' @return A data.frame object
#' @export
getContentSummaryFromDB <- function(dbPath) {
  
}
deleteSampleDataFromDB}
\usage{
deleteSampleDataFromDB(dbPath, sampleNames)
}
\arguments{
\item{dbPath}{Path to the sqlite database}

\item{sampleNames}{The names of the samples for which all relevant datasets
should be deleted from the database. Tip: Use RSQLite::dbReadTable function
to read the table 'processedSamples' to see which samples are available in
the database.}
}
\value{
SQLiteConnection object with updated contents in the dbPath
}
\description{
Given a list of sample names, the function deletes all datasets calculated
for the given samples from the database.
}
