% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAS.R
\name{extractCASFromName}
\alias{extractCASFromName}
\title{Extract CAS number from string of drug name}
\usage{
extractCASFromName(drugname)
}
\arguments{
\item{drugname}{drug/compound name to parse}
}
\value{
cas or list of cas numbers for drug/compound name
}
\description{
Extract CAS number from string of drug name
}
