% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_analysis.R
\name{findDifferentialMotifs}
\alias{findDifferentialMotifs}
\title{Find Differential Motifs}
\usage{
findDifferentialMotifs(
  querySeqs,
  controlSeqs,
  motifWidth = 6,
  motifN,
  nCores,
  maxMismatch = 1
)
}
\description{
Find Differential Motifs
}
