% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_analysis.R
\name{generateKmers}
\alias{generateKmers}
\title{Generate K-mers}
\usage{
generateKmers(k, letters = c("A", "C", "G", "T"))
}
\arguments{
\item{k}{The length of the strings to be generated}

\item{letters}{A character vector}
}
\value{
Vector of strings
}
\description{
Given a list of characters, generates all possible fixed length strings
}
\examples{
generateKmers(3, c('A', 'C', 'G'))

}
