% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneSets}
\alias{geneSets}
\title{Random test gene sets}
\format{A list of vectors, where each list element corresponds to a (randomized)
gene set, where genes are represented by Entrez ids.}
\usage{
geneSets
}
\value{
A list object
}
\description{
This dataset contains random sets of genes with Entrez ids that is designed
to represent the data that can be parsed from MSIGDB database. 
using the \code{parseMsigdb} function.
}
\details{
Actual curated datasets must be downloaded from the MSIGDB database or any 
gene set must be represented using lists.
}
\keyword{datasets}

