% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAS.R
\name{getCAS}
\alias{getCAS}
\title{get CAS number}
\usage{
getCAS(drugname)
}
\arguments{
\item{drugname}{name of drug/compound to parse}
}
\value{
CAS or list of CAS numbers for drug/compound name
}
\description{
This function combines different functions to get CAS number for a drug/compound name from DrugBank Vocabulary, Common Chemistry API and PubChem API
}
