% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAS.R
\name{getCASFromCC}
\alias{getCASFromCC}
\title{get CAS number from Common Chemistry API}
\usage{
getCASFromCC(drug)
}
\arguments{
\item{drugname}{drug/compound name to parse}
}
\value{
cas or list of cas numbers for drug/compound name
}
\description{
get CAS number from Common Chemistry API
}
