% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAS.R
\name{getCASFromDB}
\alias{getCASFromDB}
\title{get CAS number from drugbank}
\usage{
getCASFromDB(drugname)
}
\arguments{
\item{drugname}{drug/compound name to parse}
}
\value{
cas or list of cas numbers for drug/compound name
}
\description{
get CAS number from drugbank
}
