% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAS.R
\name{getCASFromPubChem2}
\alias{getCASFromPubChem2}
\title{get CAS number from Pubchem where listed on pubchem as substance}
\usage{
getCASFromPubChem2(drug)
}
\arguments{
\item{drugname}{drug/compound name to parse}
}
\value{
cas or list of cas numbers for drug/compound name
}
\description{
get CAS number from Pubchem where listed on pubchem as substance
}
