% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAS.R
\name{getCASNonDB}
\alias{getCASNonDB}
\title{get CAS number from sources other than drugbank - common chemistry API and pubchem API}
\usage{
getCASNonDB(drugname)
}
\arguments{
\item{drugname}{drug/compound name to parse}
}
\value{
cas or list of cas numbers for drug/compound name
}
\description{
get CAS number from sources other than drugbank - common chemistry API and pubchem API
}
