% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_functions.R
\name{getOverlapSimilarityMatrix}
\alias{getOverlapSimilarityMatrix}
\title{getOverlapSimilarityMatrix}
\usage{
getOverlapSimilarityMatrix(queryRegionsList, nodeN = 2, ...)
}
\arguments{
\item{queryRegionsList}{A GRangesList object}

\item{nodeN}{(default: 2) Number of nodes to use to parallellize the process}

\item{...}{Other arguments passed to IRanges::overlapsAny function}
}
\value{
A matrix with equal number of rows and columns where the size is 
  determined by the number of list elements in the input object. Matrix 
  values are between 0 and 1, where 0 means no similarity (no overlap) and 1 
  means all elements of object A overlap have at least one overlap with some 
  element in B (but not necessarily vice versa).
}
\description{
This function takes as input a GRangesList object and finds pairwise
overlapping intervals in all list elements. Then, it creates a similarity
matrix based on the percentage of intervals in each GRanges object to overlap
any intervals in every other GRanges object in the list.
}
\examples{
data(queryRegions)
queryRegionsList <- GRangesList(queryRegions, queryRegions, 
                                queryRegions, queryRegions)
getOverlapSimilarityMatrix(queryRegionsList, type = 'any')  
}
