% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_analysis.R
\name{getPWM}
\alias{getPWM}
\title{getPWM}
\usage{
getPWM(sequences, letters = c("A", "C", "G", "T"))
}
\arguments{
\item{sequences}{vector of strings of equal widths.}

\item{letters}{vector of characters to consider as the an alphabet}
}
\value{
A matrix of position-specific-weights.
}
\description{
Given a vector of strings of equal width, generate a position-specific 
weight matrix based on the frequency of occurrence of the unique letters
found in the sequences.
}
\examples{

sequences = c("GGAGAG", "GAAGAA", "TGAGAA", "GGAGAA", "GAAGAA")
getPWM(sequences)

}
