% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{getTxdbFeatures}
\alias{getTxdbFeatures}
\title{getTxdbFeatures}
\usage{
getTxdbFeatures(txdb)
}
\arguments{
\item{txdb}{A txdb object imported by GenomicFeatures::makeTxDb family of
functions}
}
\value{
A list of GRanges objects
}
\description{
This function takes as input a txdb object from GenomicFeatures library. Then
extracts the coordinates of gene features such as promoters, introns, exons,
intron-exon boundaries, 5'/3' UTRs, and whole transcripts.
}
\examples{
data(gff)
txdb <- GenomicFeatures::makeTxDbFromGRanges(gff)
txdbFeatures <- getTxdbFeatures(txdb)

}

