% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{getTxdbFeaturesFromGff}
\alias{getTxdbFeaturesFromGff}
\title{getTxdbFeaturesFromGff}
\usage{
getTxdbFeaturesFromGff(gff)
}
\arguments{
\item{gff}{A GRanges object imported by \code{importGtf} function}
}
\value{
A list of GRanges objects
}
\description{
This function takes as input a GRanges object from the output of
\code{importGtf} function. Then extracts the coordinates of gene features
such as promoters, introns, exons, intron-exon boundaries, 5'/3' UTRs and
whole transcripts.
}
\examples{
gff <- importGtf(file='annotation.gtf')
txdbFeatures <- getTxdbFeaturesFromGff(gff)

}

