% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{getTxdbFeaturesFromGRanges}
\alias{getTxdbFeaturesFromGRanges}
\title{getTxdbFeaturesFromGRanges}
\usage{
getTxdbFeaturesFromGRanges(gffData)
}
\arguments{
\item{gffData}{A GRanges object imported by \code{importGtf} function}
}
\value{
A list of GRanges objects
}
\description{
This function takes as input a GRanges object that contains GTF file contents
(e.g from the output of \code{importGtf} function). 
Then extracts the coordinates of gene features
such as promoters, introns, exons, 5'/3' UTRs and
whole transcripts.
}
\examples{
data(gff)
txdbFeatures <- getTxdbFeaturesFromGRanges(gffData = gff)

}

