% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gff}
\alias{gff}
\title{Sample GTF file imported as a GRanges object}
\format{GRanges object with 238010 ranges and 16 metadata columns}
\source{
\url{ftp://ftp.ensembl.org/pub/release-75/gtf/homo_sapiens/Homo_sapiens.GRCh37.75.gtf.gz}
}
\usage{
gff
}
\description{
This dataset contains genomic annotation data from Ensembl version 75 for Homo sapiens
downloaded from Ensembl.
The gtf file is imported via the \code{importGtf} function and a subset of the data
is selected by choosing features found on 'chr1'.
}
\keyword{datasets}

