% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigDB}
\alias{msigDB}
\title{MSIGDB gene sets}
\format{A list of vectors, where each list element corresponds to a
curated pathways gene set, where genes are represented by Entrez ids.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb}
}
\usage{
msigDB
}
\description{
This dataset contains curated sets of genes with Entrez ids from Molecular Signatures Database.
The data is imported using the \code{parseMsigdb} function.
}
\keyword{datasets}

