% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{printMsigdbDataset}
\alias{printMsigdbDataset}
\title{Print MSIGDB Dataset to a file}
\usage{
printMsigdbDataset()
}
\description{
This function is deprecated. For functional enrichment 
analysis, use findEnrichedFunctions.
}
