% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queryRegions}
\alias{queryRegions}
\title{Sample BED file imported as a GRanges object}
\format{GRanges object with 10000 ranges and 2 metadata columns}
\source{
\url{http://dorina.mdc-berlin.de/regulators}
}
\usage{
queryRegions
}
\description{
This dataset contains a randomly selected sample of human LIN28A protein binding
sites detected by HITS-CLIP analysis downloaded from DoRina database (LIN28A HITS-CLIP hESCs (Wilbert 2012)).
The BED file is imported via the \code{importBed} function and a subset of the data
is selected by randomly choosing 10000 regions.
}
\keyword{datasets}

