% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{retrieveOrthologs}
\alias{retrieveOrthologs}
\title{retrieveOrthologs}
\usage{
retrieveOrthologs()
}
\description{
This function is deprecated. For functional enrichment 
analysis, use findEnrichedFunctions.
}
