% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{runGSEA}
\alias{runGSEA}
\title{runGSEA}
\usage{
runGSEA(geneSetList, species = "human", backgroundGenes, targetedGenes)
}
\arguments{
\item{geneSetList}{A named list of vectors where each vector consists of a
set of Entrez gene ids (for instance, returned by \code{parseMsigdb}
function)}

\item{species}{A character string denoting the species under analysis. 
Options are 'human', 'mouse', 'fly' and 'worm'.}

\item{backgroundGenes}{A vector of Ensembl gene ids that serve as background 
set of genes for GO term enrichment. In the context of RCAS, this should be
the whole set of genes found in an input GTF file.}

\item{targetedGenes}{A vector of Ensembl gene ids that serve as the set for 
which GSEA should be carried out. In the context of RCAS, this should be
the set of genes that overlap the query regions}
}
\value{
A data.frame object containing enriched gene sets and associated
  statistics
}
\description{
This function is used to facilitate gene set enrichment analysis (GSEA) for a
given set of genes
}
\examples{

#load test data
data(geneSets)
data(gff)
data(queryRegions)
#get all genes from the gff data
backgroundGenes <- unique(gff$gene_id)
#get genes that overlap query regions
overlaps <- queryGff(queryRegions, gff)
targetedGenes <- unique(overlaps$gene_id)
resultsGSEA <- runGSEA(geneSetList = geneSets,
          species = 'human',
          backgroundGenes = backgroundGenes,
          targetedGenes = targetedGenes)
}
