% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseSetUpFunctions.R
\name{summarizeDatabaseContent}
\alias{summarizeDatabaseContent}
\title{summarizeDatabaseContent}
\usage{
summarizeDatabaseContent(dbPath)
}
\arguments{
\item{dbPath}{Path to the sqlite database}
}
\value{
A data.frame object
}
\description{
Given a path to an sqlite database created using RCAS::createDB function,
accesses the database and provides a quick summary of available samples 
and number of entries of each sample in the available tables of the 
database.
}
