% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{summarizeQueryRegions}
\alias{summarizeQueryRegions}
\title{summarizeQueryRegions}
\usage{
summarizeQueryRegions(queryRegions, txdbFeatures)
}
\arguments{
\item{queryRegions}{GRanges object imported from a BED file using}

\item{txdbFeatures}{List of GRanges objects - output of \code{getTxdbFeaturesFromGff} and \code{getTxdbFeatures} functions}
}
\value{
A data frame with two columns where first column holds the features and second column holds the counts
}
\description{
This function is used to count how many query regions overlap with different types of gene features
}
\examples{

peaks <- importBed('input.BED')
gff <- importGtf('annotation.gtf')
txdbFeatures <- getTxdbFeaturesFromGff(gff)
summary <- summarizeQueryRegions(queryRegions=peaks, txdbFeatures=txdbFeatures)

}

