% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{calculateCoverageProfileFromTxdb}
\alias{calculateCoverageProfileFromTxdb}
\title{calculateCoverageProfileFromTxdb}
\usage{
calculateCoverageProfileFromTxdb()
}
\description{
This function is deprecated. Use ?calculateCoverageProfile instead.
}
