% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{calculateCoverageProfileListFromTxdb}
\alias{calculateCoverageProfileListFromTxdb}
\title{calculateCoverageProfileListFromTxdb}
\usage{
calculateCoverageProfileListFromTxdb()
}
\description{
This function is deprecated. Use ?calculateCoverageProfileList instead.
}
