% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseSetUpFunctions.R
\name{createDB}
\alias{createDB}
\title{createDB}
\usage{
createDB(dbPath = file.path(getwd(), "rcasDB.sqlite"), projDataFile,
  gtfFilePath, update = FALSE, nodeN = 1)
}
\arguments{
\item{dbPath}{Path to the sqlite database file (could be an existing file or 
a new file path to be created at the given path)}

\item{projDataFile}{A file consisting of meta-data about the input samples. 
Must minimally consist of two columns: 1. sampleName (name of the sample) 
2. bedFilePath (full path to the location of the BED file containing data 
for the sample)}

\item{gtfFilePath}{Path to the GTF file (preferably downloaded from the 
Ensembl database) that contains genome annotations}

\item{update}{TRUE/FALSE (default: FALSE) whether an existing database 
should be updated}

\item{nodeN}{Number of cpus to use for parallel processing (default: 1)}
}
\value{
Path to an SQLiteConnection object created by RSQLite package
}
\description{
Creates and sqlite database consisting of various tables of data obtained 
from processed BED files
}
