% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_analysis.R
\name{getMotifSummaryTable}
\alias{getMotifSummaryTable}
\title{getMotifSummaryTable}
\usage{
getMotifSummaryTable(motifResults)
}
\arguments{
\item{motifResults}{Output object of \code{runMotifRG} function}
}
\value{
A data.frame object containing summary statistics about the
  discovered motifs
}
\description{
A repurposed/simplified version of the \code{motifRG::summaryMotif} function.
}
\examples{

bed <- importBed('input.bed')
motifResults <- runMotifRG(queryRegions=bed, genomeVersion='hg19')
motifSummary <- getMotifSummaryTable(motifResults)

}

