% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_analysis.R
\name{getMotifSummaryTable}
\alias{getMotifSummaryTable}
\title{getMotifSummaryTable}
\usage{
getMotifSummaryTable(motifResults)
}
\arguments{
\item{motifResults}{Output object of \code{runMotifRG} function}
}
\value{
A data.frame object containing summary statistics about the
  discovered motifs
}
\description{
A repurposed/simplified version of the \code{motifRG::summaryMotif} function.
}
\examples{

data(queryRegions)
motifResults <- runMotifRG(queryRegions = queryRegions,
                          genomeVersion = 'hg19',
                          resize = 15,
                          motifN = 1,
                          nCores = 2)
motifSummary <- getMotifSummaryTable(motifResults)

}
