% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{printMsigdbDataset}
\alias{printMsigdbDataset}
\title{Print MSIGDB Dataset to a file
This function is used to print a MSIGDB dataset into a file. Mostly
useful when human data is mapped to another species, and that mapping
is required to run the report.}
\usage{
printMsigdbDataset(dataset, outputFilename)
}
\description{
Print MSIGDB Dataset to a file
This function is used to print a MSIGDB dataset into a file. Mostly
useful when human data is mapped to another species, and that mapping
is required to run the report.
}

