% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{runReport}
\alias{runReport}
\title{Generate a RCAS Report for a CLIP-Seq-based experiment}
\usage{
runReport(queryFilePath = "testdata", gffFilePath = "testdata",
  msigdbFilePath = "testdata", annotationSummary = TRUE,
  goAnalysis = TRUE, msigdbAnalysis = TRUE, motifAnalysis = TRUE,
  genomeVersion = "hg19", outDir = getwd())
}
\arguments{
\item{queryFilePath}{a BED format file which contains genomic coordinates of
protein-RNA binding sites}

\item{gffFilePath}{A GTF format file which contains genome annotations
(preferably from ENSEMBL)}

\item{msigdbFilePath}{Gene set annotations for Homo sapiens from Molecular
Signatures Database}

\item{annotationSummary}{TRUE/FALSE (default: TRUE) A switch to decide if
RCAS should provide annotation summaries from overlap operations}

\item{goAnalysis}{TRUE/FALSE (default: TRUE) A switch to decide if RCAS
should run GO term enrichment analysis}

\item{msigdbAnalysis}{TRUE/FALSE (default: TRUE) A switch to decide if RCAS
should run gene set enrichment analysis}

\item{motifAnalysis}{TRUE/FALSE (default: TRUE) A switch to decide if RCAS
should run motif analysis}

\item{genomeVersion}{A character string to denote for which genome version
the analysis is being done. Available options are hg19 (human), mm9
(mouse), ce10 (worm) and dm3 (fly).}

\item{outDir}{Path to the output directory. (default: current working
directory)}
}
\value{
An html generated using rmarkdown/knitr/pandoc that contains
  interactive figures, tables, and text that provide an overview of the
  experiment
}
\description{
This is the main report generation function for RCAS. This function can take
a BED file, a GTF file and optionally an MSIGDB annotation; and use these
input to run multiple RCAS functions to create a summary report regarding the
annotation data that overlap the input BED file, enrichment analysis for GO
terms, gene sets from MSIGDB, and motif analysis.
}
\examples{
#Default run will generate a report using built-in test data for hg19 genome.
runReport()

#A custom run for human
runReport( queryFilePath = 'input.BED',
           gffFilePath = 'annotation.gtf',
           msigdbFilePath = 'human_msigdb.gmt',
# To turn off certain modules of the report
runReport( queryFilePath = 'input.BED',
           gffFilePath = 'annotation.gtf',
           msigdbFilePath = 'human_msigdb.gmt',
           motifAnalysis = FALSE,
           goAnalysis = FALSE
           )
# To run the pipeline for species other than human
# If the msigdb module is needed, the msigdbFilePath
# must be set to the MSIGDB annotations for 'human'.
# MSIGDB datasets for other species will be calculated
# in the background using the createOrthologousMsigdbDataset
# function
runReport( queryFilePath = 'input.mm9.BED',
           gffFilePath = 'annotation.mm9.gtf',
           msigdbFilePath = 'msigdb.human.gmt',
           genomeVersion = 'mm9' )

}

